Rewrite
First commit containing major rewrite. Created a sequence class that is now used by both a UI and a command line script. Also added the ability to load in a localizer to check localization. The offsets still need to be tested, as I am not sure about the direction (sign).